+++ updateStaff.php
@@ -1,0 +1,7 @@
+<?php 
+    include './db/db.php';
+    include 'helper.php';
+    if (isset($_POST["submit"])) {
+        console_log($_POST["last_name"]);
+    }
+?>
